Create README.md
Hi,
We hope you're doing well.

We wanted to sincerely apologize for missing the README file in our initial submission. Our goal was to aim for a grade 5, and while reviewing everything again, we realized that we had unfortunately overlooked this important part.

We've now added the README file and are kindly asking if you would consider accepting it at this stage. We're really sorry for the oversight and truly appreciate your understanding.

Thank you so much for your time and support!

Warm regards,
group 22